Use a <caption> for the template statistics table
This is semantically more accurate because it’s describing the whole
table, not just the first column.

Adjusting the font-size to make it sit within the ‘In the last 7 days’
section. Adjusting the spacing because now that we have more borders we
don’t need quite so much whitespace to separate different bits of the
page.