Say ‘recently’ not ‘yet’ because jobs aren’t forever
We’re going to change the jobs page so that it only shows jobs within
the last n days (your data retention)<sup>1</sup>.

This will match how long uploaded letters stick around for.

Therefore it’s not accurate to say ‘yet’, because that implies all-time.
Since the data retention for different channels could be different it’s
hard and maybe unhelpful to give an exact time period. ‘Recently’ is
content we used here before, but then changed.

1. https://www.pivotaltracker.com/story/show/171623239